+++ packages/expo-router/CHANGELOG.md
@@ -33,6 +33,7 @@
 
 ### 💡 Others
 
+- Improve expo-router test coverage. ([#47720](https://github.com/expo/expo/pull/47720) by [@Ubax](https://github.com/Ubax))
 - Rewrite native tabs using standard-navigation ([#46457](https://github.com/expo/expo/pull/46457) by [@Ubax](https://github.com/Ubax))
 - [Internal] Split `useLoaderData()` into a document cache and a per-mount Suspense store ([#47365](https://github.com/expo/expo/pull/47365) by [@hassankhan](https://github.com/hassankhan))